design of register form and basic functions that do nothing
