+++ client/src/assets/scss/views/_homepage.scss
@@ -205,7 +205,7 @@
     }
 
     @include for-wide-layout {
-        width: 70%;
+        width: 80%;
         padding-bottom: 30px;
     }
