+++ client/src/views/home-page.jsx
@@ -1,23 +1,10 @@
-import { useState } from "react"
-import { Outlet, useLocation, useNavigate, Link } from "react-router-dom"
+import { useLocation, useNavigate } from "react-router-dom"
+import { LoginSignup } from "./auth/LoginSignup"
 
 export function HomePage() {
-    const navigate = useNavigate()
     const location = useLocation()
-    const [userDetails, setUserDetails] = useState({})
-
-    const handleChange = ({ target }) => {
-        const { name, value } = target
-        setUserDetails({ ...userDetails, [name]: value })
-    }
-
-    console.log('userDetails:', userDetails);
+    const navigate = useNavigate()
     
-
-    const onChangeForm = () => {
-        location.pathname === '/login' ? navigate('/signup') : navigate('/login')
-    }
-
 
     return <div className="home-page ">
         <div className="dark-fade-div">
@@ -45,33 +32,7 @@ export function HomePage() {
                         </button>
                     </section>}
                     {/* Login / Signup */}
-                    {location.pathname !== '/' && <section className="register-modal">
-                        <form className="register-modal__form">
-                            <input className="register-modal__input" name="username" onChange={handleChange} type="text" placeholder="הכניסו שם" />
-                            {location.pathname === '/signup' && <> <input className="register-modal__input" name="phone" onChange={handleChange} type="text" placeholder="הכניסו פלאפון" />
-                            <input className="register-modal__input" name="mail" onChange={handleChange} type="text" placeholder="הכניסו מייל" /></>}
-                            <input className="register-modal__input" name="password" onChange={handleChange} type="text" placeholder="הכניסו סיסמא" />
-                            <div className="password-options">
-                                <div className="password-options__remmember-user">
-                                    <input className="password-options__input" type="checkbox" name="" id="" />
-                                    <label>זכור אותי</label>
-                                </div>
-                                <Link to={'/'} className='password-options__forget-password-link'>שכחת סיסמא?</Link>
-                            </div>
-                            <button className="register-modal__cta-btn cta-btn">התחבר</button>
-                        </form>
-                        <div className="register-modal__alternatives-container">
-                            <div className="register-modal__or-seperator">
-                                <div className="border"></div>
-                                <p>או</p>
-                                <div className="border"></div>
-                            </div>
-                            <button className="register-modal__facebook-btn">התחבר באמצעות פייסבוק</button>
-                            <button className="register-modal__google-btn">התחבר באמצעות גוגל</button>
-                            <button className="register-modal__alternative-btn" onClick={onChangeForm}>{location.pathname === '/signup' ? 'התחבר עם חשבון קיים' : 'הירשם'}</button>
-                            <Link to={'/'} className='register-modal__skip-link'>דלגו לבינתיים </Link>
-                        </div>
-                    </section>}
+                    {location.pathname !== '/' && <LoginSignup />}
                 </section>
             </div>
         </div>
